histogram: add TimeHistogram
This change pulls in the TimeHistogram from https://github.com/Noah-Kennedy/prometheus-client-utils.git, allowing us to get rid of tha crate, and changes the HistogramTimer to use that histogram instead of the one from the prometheus_client crate.

It also makes a minor change to CI, adding a new job which runs `cargo check` on each feature.