Backport new Family methods from latest prometheus-client
`prometheus-client 0.24` exposes methods to remove individual metrics
from a `Family` as well as clearing the family entirely. These are
useful, but were not implemented in the 0.18 release that we are stuck
with. To backport them, we need to implement `Family` ourselves.

Thankfully, the code for `Family` is not very complex. I copied the
implementations for all existing methods and added `remove(label_set)`
and `clear()`.